Initial renaming of source folders
