+++ noa-harvester/generate_netrc_local.sh
